Move sanitize_value to be common Fix Indian Passport Number regex
