+++ presidio-analyzer/presidio_analyzer/__init__.py
@@ -15,7 +15,6 @@ from presidio_analyzer.recognizer_registry import RecognizerRegistry
 from presidio_analyzer.analyzer_engine import AnalyzerEngine
 from presidio_analyzer.batch_analyzer_engine import BatchAnalyzerEngine
 from presidio_analyzer.analyzer_request import AnalyzerRequest
-from presidio_analyzer.analyzer_utils import PresidioAnalyzerUtils
 from presidio_analyzer.context_aware_enhancers import ContextAwareEnhancer
 from presidio_analyzer.context_aware_enhancers import LemmaContextAwareEnhancer
 from presidio_analyzer.analyzer_engine_provider import AnalyzerEngineProvider
@@ -51,6 +50,5 @@ __all__ = [
     "ContextAwareEnhancer",
     "LemmaContextAwareEnhancer",
     "BatchAnalyzerEngine",
-    "PresidioAnalyzerUtils",
     "AnalyzerEngineProvider",
 ]
