+++ presidio-analyzer/presidio_analyzer/predefined_recognizers/au_abn_recognizer.py
@@ -1,5 +1,7 @@
 from typing import List, Optional, Tuple
 
+from validation.validation_utils import ValidationUtils
+
 from presidio_analyzer import Pattern, PatternRecognizer
 
 
@@ -72,7 +74,7 @@ class AuAbnRecognizer(PatternRecognizer):
         :return: A bool indicating whether the validation was successful.
         """
         # Pre-processing before validation checks
-        text = self.__sanitize_value(pattern_text, self.replacement_pairs)
+        text = ValidationUtils.sanitize_value(pattern_text, self.replacement_pairs)
         abn_list = [int(digit) for digit in text if not digit.isspace()]
 
         # Set weights based on digit position
@@ -85,9 +87,3 @@ class AuAbnRecognizer(PatternRecognizer):
             sum_product += abn_list[i] * weight[i]
         remainder = sum_product % 89
         return remainder == 0
-
-    @staticmethod
-    def __sanitize_value(text: str, replacement_pairs: List[Tuple[str, str]]) -> str:
-        for search_string, replacement_string in replacement_pairs:
-            text = text.replace(search_string, replacement_string)
-        return text
