+++ presidio-analyzer/presidio_analyzer/predefined_recognizers/it_vat_code.py
@@ -1,5 +1,7 @@
 from typing import List, Optional, Tuple
 
+from validation.validation_utils import ValidationUtils
+
 from presidio_analyzer import Pattern, PatternRecognizer
 
 
@@ -62,7 +64,7 @@ class ItVatCodeRecognizer(PatternRecognizer):
         """
 
         # Pre-processing before validation checks
-        text = self.__sanitize_value(pattern_text, self.replacement_pairs)
+        text = ValidationUtils.sanitize_value(pattern_text, self.replacement_pairs)
 
         # Edge-case that passes the checksum even though it is not a
         # valid italian vat code.
@@ -88,9 +90,3 @@ class ItVatCodeRecognizer(PatternRecognizer):
             result = False
 
         return result
-
-    @staticmethod
-    def __sanitize_value(text: str, replacement_pairs: List[Tuple[str, str]]) -> str:
-        for search_string, replacement_string in replacement_pairs:
-            text = text.replace(search_string, replacement_string)
-        return text
