+++ presidio-analyzer/presidio_analyzer/predefined_recognizers/medical_license_recognizer.py
@@ -1,5 +1,7 @@
 from typing import List, Optional, Tuple
 
+from validation.validation_utils import ValidationUtils
+
 from presidio_analyzer import Pattern, PatternRecognizer
 
 # https://www.meditec.com/blog/dea-numbers-what-do-they-mean
@@ -50,7 +52,9 @@ class MedicalLicenseRecognizer(PatternRecognizer):
         )
 
     def validate_result(self, pattern_text: str) -> bool:  # noqa D102
-        sanitized_value = self.__sanitize_value(pattern_text, self.replacement_pairs)
+        sanitized_value = ValidationUtils.sanitize_value(
+            pattern_text, self.replacement_pairs
+        )
         checksum = self.__luhn_checksum(sanitized_value)
 
         return checksum
@@ -67,9 +71,3 @@ class MedicalLicenseRecognizer(PatternRecognizer):
         checksum *= -1
         checksum += 2 * sum(even_digits) + sum(odd_digits)
         return checksum % 10 == 0
-
-    @staticmethod
-    def __sanitize_value(text: str, replacement_pairs: List[Tuple[str, str]]) -> str:
-        for search_string, replacement_string in replacement_pairs:
-            text = text.replace(search_string, replacement_string)
-        return text
