+++ presidio-analyzer/tests/test_in_aadhaar_recognizer.py
@@ -45,3 +45,47 @@ def test_when_aadhaar_in_text_then_all_aadhaars_found(
             expected_position[1],
             expected_score,
         )
+
+palindrome_test_set = [
+    ["abMA", False, False],
+    ["abCba", False, True],
+    ["ABBA", False, True],
+    ["aBba", True, True],
+    ["NotAPalindrome", True, False],
+]
+
+
+@pytest.mark.parametrize(
+    "input_text,case_sensitive, expected_output", palindrome_test_set
+)
+def test_is_palindrome(input_text, case_sensitive, expected_output):
+    """
+    Test if input is a true palindrome as defined in base class.
+
+    :param input_text: input text to validate
+    :param case_sensitive: flag to calculate palindrome with no case
+    :param expected_output: calculated output
+    :return: True/False
+    """
+    assert (
+            InAadhaarRecognizer._is_palindrome(input_text, case_sensitive)
+            == expected_output
+    )
+
+
+verhoeff_test_set = [
+    [312345678909, True],
+    [400123456787, True],
+    [123456789012, False],
+]
+
+@pytest.mark.parametrize("input_number, is_verhoeff", verhoeff_test_set)
+def test_is_verhoeff(input_number, is_verhoeff):
+    """
+    Test to assert verhoeff number validation based on checksum from base class.
+
+    :param input_number: input integer
+    :param is_verhoeff: expected flag
+    :return: True/False
+    """
+    assert InAadhaarRecognizer._is_verhoeff_number(input_number) == is_verhoeff
